changed frn datatype to integer from bigint

+++ app/data/models/delinkedCalculation.js
@@ -5,7 +5,7 @@ module.exports = (sequelize, DataTypes) => {
     applicationId: { type: DataTypes.INTEGER, allowNull: false },
     calculationId: { type: DataTypes.INTEGER, primaryKey: true, allowNull: false },
     sbi: { type: DataTypes.INTEGER, allowNull: false },
-    frn: { type: DataTypes.BIGINT, allowNull: false },
+    frn: { type: DataTypes.INTEGER, allowNull: false },
     paymentBand1: { type: DataTypes.INTEGER, allowNull: false },
     paymentBand2: { type: DataTypes.INTEGER, allowNull: false },
     paymentBand3: { type: DataTypes.INTEGER, allowNull: false },
